Grammar fix in index.html
"How it works" had a question mark at the end in two places. Removed the question mark so the grammar is correct and consistent with the navbar entry "How It Works". Could also be "How does it work?" but this way the least is changed. https://ell.stackexchange.com/questions/17778/how-it-works-vs-how-does-it-work for grammar rule.